Move generic Card definition to top-level package
